app_t2/requirements.txt
streamlit==1.28.1
opencv-python==4.8.1.78
ultralytics==8.0.196
numpy==1.24.3
Pillow==10.0.1
torch>=1.13.0
torchvision>=0.14.0